+++ src/molecules/SliderItem.js
